refactor(web): extract 2 components from Library and BookDedup (STRUCT-9 part 1)
Extracts:
- LibrarySoftDeletedSection from Library.tsx
- DedupBookTab from BookDedup.tsx
- dedupHelpers.tsx with shared utilities

Library.tsx: 3333→3243 lines (-90)
BookDedup.tsx: 3656→3424 lines (-232)

No logic or state changes. Structure audit STRUCT-9 (partial).
