+++ web/src/pages/BookDedup.tsx
@@ -1,5 +1,5 @@
 // file: web/src/pages/BookDedup.tsx
-// version: 3.17.0
+// version: 3.18.0
 // guid: c3d4e5f6-a7b8-9c0d-1e2f-book0dedup02
 
 import { useState, useEffect, useCallback, useMemo } from 'react';
@@ -22,8 +22,6 @@ import {
   CardContent,
   CardActions,
   Stack,
-  Radio,
-  RadioGroup,
   FormControlLabel,
   Tab,
   Tabs,
@@ -68,6 +66,7 @@ import MicIcon from '@mui/icons-material/Mic';
 import BusinessIcon from '@mui/icons-material/Business';
 import CleaningServicesIcon from '@mui/icons-material/CleaningServices';
 import type { SuggestionRoles } from '../services/api';
+import { DedupBookTab } from '../components/dedup/DedupBookTab';
 
 /** Strip "(Unabridged)", "(Abridged)", and leading "[Series X]" from display titles */
 function cleanDisplayTitle(title: string): string {
@@ -291,229 +290,7 @@ function PaginationControls({ total, page, rowsPerPage, onPageChange, onRowsPerP
 }
 
 // ---- Book Dedup Tab ----
-function BookDedupTab() {
-  const [groups, setGroups] = useState<Book[][]>([]);
-  const [totalDuplicates, setTotalDuplicates] = useState(0);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [activeOp, setActiveOp] = useState<Operation | null>(null);
-  const [mergeSuccess, setMergeSuccess] = useState<string | null>(null);
-  const [keepSelections, setKeepSelections] = useState<Record<string, string>>({});
-  const [selectedGroups, setSelectedGroups] = useState<Set<string>>(new Set());
-  const [confirmOpen, setConfirmOpen] = useState(false);
-  const pagination = usePagination(groups.length);
-
-  const fetchDuplicates = useCallback(async () => {
-    setLoading(true);
-    setError(null);
-    try {
-      const data = await api.getBookDuplicates();
-      setGroups(data.groups || []);
-      setTotalDuplicates(data.duplicate_count || 0);
-      const defaults: Record<string, string> = {};
-      (data.groups || []).forEach((g, i) => {
-        if (g.length > 0) defaults[`group-${i}`] = g[0].id;
-      });
-      setKeepSelections(defaults);
-      setSelectedGroups(new Set());
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Failed to fetch duplicates');
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  useEffect(() => { fetchDuplicates(); }, [fetchDuplicates]);
-
-  const handleMerge = async (group: Book[], groupKey: string) => {
-    const keepId = keepSelections[groupKey];
-    if (!keepId) return;
-    const mergeIds = group.filter((b) => b.id !== keepId).map((b) => b.id);
-    setMergeSuccess(null);
-    await runOperationWithPolling(
-      () => api.mergeBooks(keepId, mergeIds),
-      setActiveOp,
-      (final) => {
-        if (final.status === 'failed') {
-          setError(final.error_message || 'Merge failed');
-        } else {
-          setMergeSuccess(`Merged duplicates of "${group[0]?.title}"`);
-          setGroups((prev) => prev.filter((_, i) => `group-${i}` !== groupKey));
-          setSelectedGroups((prev) => { const next = new Set(prev); next.delete(groupKey); return next; });
-        }
-      },
-      setError,
-    );
-  };
-
-  const handleMergeSelected = async () => {
-    setMergeSuccess(null);
-    for (let i = 0; i < groups.length; i++) {
-      const groupKey = `group-${i}`;
-      if (!selectedGroups.has(groupKey)) continue;
-      const group = groups[i];
-      const keepId = keepSelections[groupKey];
-      if (!keepId) continue;
-      const mergeIds = group.filter((b) => b.id !== keepId).map((b) => b.id);
-      try {
-        const initial = await api.mergeBooks(keepId, mergeIds);
-        setActiveOp(initial);
-        await api.pollOperation(initial.id, (update) => setActiveOp(update));
-      } catch (err) {
-        setError(err instanceof Error ? err.message : `Failed to merge "${group[0]?.title}"`);
-      }
-    }
-    setActiveOp(null);
-    setMergeSuccess(`Merged ${selectedGroups.size} selected group(s)`);
-    fetchDuplicates();
-  };
-
-  const handleMergeAll = async () => {
-    setConfirmOpen(false);
-    setMergeSuccess(null);
-    for (let i = 0; i < groups.length; i++) {
-      const group = groups[i];
-      const groupKey = `group-${i}`;
-      const keepId = keepSelections[groupKey];
-      if (!keepId) continue;
-      const mergeIds = group.filter((b) => b.id !== keepId).map((b) => b.id);
-      try {
-        const initial = await api.mergeBooks(keepId, mergeIds);
-        setActiveOp(initial);
-        await api.pollOperation(initial.id, (update) => setActiveOp(update));
-      } catch (err) {
-        setError(err instanceof Error ? err.message : `Failed to merge "${group[0]?.title}"`);
-      }
-    }
-    setActiveOp(null);
-    setMergeSuccess('Merged all duplicate books');
-    fetchDuplicates();
-  };
-
-  const toggleGroup = (key: string) => {
-    setSelectedGroups((prev) => {
-      const next = new Set(prev);
-      if (next.has(key)) next.delete(key); else next.add(key);
-      return next;
-    });
-  };
-
-  const toggleAll = () => {
-    if (selectedGroups.size === groups.length) {
-      setSelectedGroups(new Set());
-    } else {
-      setSelectedGroups(new Set(groups.map((_, i) => `group-${i}`)));
-    }
-  };
-
-  const busy = activeOp !== null;
-
-  return (
-    <Box>
-      <Box sx={{ display: 'flex', alignItems: 'center', mb: 2 }}>
-        <Typography variant="body2" color="text.secondary" sx={{ flexGrow: 1 }}>
-          Detects books with identical titles and authors at different file paths.
-        </Typography>
-        <Stack direction="row" spacing={1}>
-          {groups.length > 0 && (
-            <>
-              <Button size="small" onClick={toggleAll} disabled={busy}>
-                {selectedGroups.size === groups.length ? 'Deselect All' : 'Select All'}
-              </Button>
-              {selectedGroups.size > 0 && (
-                <Button variant="contained" color="primary" startIcon={<MergeIcon />}
-                  onClick={handleMergeSelected} disabled={busy}>
-                  Merge Selected ({selectedGroups.size})
-                </Button>
-              )}
-              <Button variant="contained" color="warning" startIcon={<MergeIcon />}
-                onClick={() => setConfirmOpen(true)} disabled={busy}>
-                Merge All ({totalDuplicates})
-              </Button>
-            </>
-          )}
-          <Tooltip title="Refresh"><IconButton onClick={fetchDuplicates} disabled={loading || busy}><RefreshIcon /></IconButton></Tooltip>
-        </Stack>
-      </Box>
-
-      <OperationProgress operation={activeOp} />
-      {error && <Alert severity="error" sx={{ mb: 2 }} onClose={() => setError(null)}>{error}</Alert>}
-      {mergeSuccess && <Alert severity="success" sx={{ mb: 2 }} icon={<CheckCircleIcon />} onClose={() => setMergeSuccess(null)}>{mergeSuccess}</Alert>}
-
-      {loading ? (
-        <Box sx={{ display: 'flex', justifyContent: 'center', py: 4 }}><CircularProgress /></Box>
-      ) : groups.length === 0 ? (
-        <Paper sx={{ p: 4, textAlign: 'center' }}>
-          <CheckCircleIcon sx={{ fontSize: 48, color: 'success.main', mb: 1 }} />
-          <Typography variant="h6">No duplicate books found</Typography>
-        </Paper>
-      ) : (
-        <>
-        <PaginationControls total={groups.length} page={pagination.page} rowsPerPage={pagination.rowsPerPage}
-          onPageChange={pagination.setPage} onRowsPerPageChange={pagination.setRowsPerPage} />
-        <Stack spacing={2}>
-          {groups.slice(pagination.startIdx, pagination.endIdx).map((group, sliceIdx) => {
-            const idx = pagination.startIdx + sliceIdx;
-            const groupKey = `group-${idx}`;
-            return (
-              <Card key={groupKey} variant="outlined">
-                <CardContent>
-                  <Box sx={{ display: 'flex', alignItems: 'center', gap: 1, mb: 1 }}>
-                    <Checkbox
-                      checked={selectedGroups.has(groupKey)}
-                      onChange={() => toggleGroup(groupKey)}
-                      disabled={busy}
-                      size="small"
-                    />
-                    <Typography variant="subtitle1" fontWeight="bold">{cleanDisplayTitle(group[0]?.title || 'Unknown')}</Typography>
-                    <Chip label={`${group.length} copies`} size="small" color="warning" variant="outlined" />
-                  </Box>
-                  <Divider sx={{ my: 1 }} />
-                  <RadioGroup value={keepSelections[groupKey] || ''}
-                    onChange={(e) => setKeepSelections((prev) => ({ ...prev, [groupKey]: e.target.value }))}>
-                    {group.map((book) => (
-                      <FormControlLabel key={book.id} value={book.id} control={<Radio size="small" />}
-                        label={
-                          <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                            <FolderIcon fontSize="small" color="action" />
-                            <Typography variant="body2" sx={{ fontFamily: 'monospace', fontSize: '0.8rem' }}>{book.file_path}</Typography>
-                            {book.itunes_persistent_id && <Chip label="iTunes" size="small" color="info" variant="outlined" />}
-                            {book.format && <Chip label={book.format} size="small" variant="outlined" />}
-                          </Box>
-                        } />
-                    ))}
-                  </RadioGroup>
-                </CardContent>
-                <CardActions>
-                  <Button startIcon={<MergeIcon />} variant="contained" size="small"
-                    onClick={() => handleMerge(group, groupKey)} disabled={busy}>
-                    Merge
-                  </Button>
-                </CardActions>
-              </Card>
-            );
-          })}
-        </Stack>
-        <PaginationControls total={groups.length} page={pagination.page} rowsPerPage={pagination.rowsPerPage}
-          onPageChange={pagination.setPage} onRowsPerPageChange={pagination.setRowsPerPage} />
-        </>
-      )}
-
-      <Dialog open={confirmOpen} onClose={() => setConfirmOpen(false)}>
-        <DialogTitle>Confirm Merge All</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            This will merge {groups.length} groups. This action cannot be undone. Are you sure?
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={() => setConfirmOpen(false)}>Cancel</Button>
-          <Button onClick={handleMergeAll} color="warning" variant="contained">Confirm</Button>
-        </DialogActions>
-      </Dialog>
-    </Box>
-  );
-}
+// Moved to web/src/components/dedup/DedupBookTab.tsx
 
 // ---- Advanced Book Dedup Scan Tab ----
 function BookDedupScanTab() {
@@ -3644,7 +3421,7 @@ export function BookDedup() {
         <Tab icon={<Badge color="default"><FingerprintIcon /></Badge>} label="Embedding" iconPosition="start" />
       </Tabs>
 
-      {tab === 0 && <BookDedupTab />}
+      {tab === 0 && <DedupBookTab />}
       {tab === 1 && <BookDedupScanTab />}
       {tab === 2 && <AuthorDedupTab />}
       {tab === 3 && <SeriesDedupTab />}
